requirements_2.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.0.0
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
yarl==1.8.1
